+++ R/carstats.R
@@ -1,8 +1,0 @@
-#' @title Data from carstats
-#' 
-#' @description
-#'   Copy from the carstats sample dataset available in Cornerstone.
-#' 
-#' @format
-#'   A \code{\link{data.table}} object with 406 rows and 9 columns.
-"carstats"