+++ R/zzz.R
@@ -1,3 +1,4 @@
 #' @import checkmate
 #' @import data.table
+#' @importFrom utils head tail
 "_PACKAGE"